DYNAMO_CHUNK_SIZE multiple bug fix
Restoring fails if the json contains a number of records which is an exact multiple of DYNAMO_CHUNK_SIZE, because the download "close" contains no records but it was assumed it did. This created an empty batch. Added a check to only create the final batch if there actually were records in the close.